cleanup & comments
