+++ apps/dotcom/sync-worker/src/replicator/getResumeType.ts
@@ -1,7 +1,8 @@
 import { SqlStorage } from '@cloudflare/workers-types'
 import { Logger } from '../Logger'
 import { ZReplicationEventWithoutSequenceInfo } from '../UserDataSyncer'
-import { CatchUpChangeCollator, parseSubscriptions } from './ChangeCollator'
+import { CatchUpChangeCollator } from './ChangeCollator'
+import { parseSubscriptions } from './Subscription'
 import { ChangeV2, Topic } from './replicatorTypes'
 
 export function getResumeType({
