+++ apps/dotcom/sync-worker/src/replicator/replicatorMigrations.test.ts
@@ -2,7 +2,7 @@ import { unlinkSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { DatabaseSync } from 'node:sqlite'
-import { parseSubscriptions } from './ChangeCollator'
+import { parseSubscriptions } from './Subscription'
 import { migrate } from './replicatorMigrations'
 import { ChangeV1 } from './replicatorTypes'
 import { MockLogger, SqlStorageAdapter } from './test-helpers'
